14409 Extract charges error to new file
