+++ Yosemite/Yosemite/Stores/CardPresentPaymentStore.swift
@@ -3,8 +3,6 @@ import Hardware
 import Networking
 import Combine
 
-
-
 // MARK: CardPresentPaymentStore
 ///
 public final class CardPresentPaymentStore: Store {
@@ -731,97 +729,3 @@ public protocol CardReaderCapableRemote {
 extension WCPayRemote: CardReaderCapableRemote {}
 extension StripeRemote: CardReaderCapableRemote {}
 
-// MARK: - WCPayChargesError
-public enum WCPayChargesError: Error, LocalizedError {
-    case noSuchChargeError(message: String)
-    case otherError(error: AnyError)
-
-    init(underlyingError error: Error) {
-        guard let errorDetails = Self.unwrapError(error: error) else {
-            self = .otherError(error: error.toAnyError)
-            return
-        }
-
-        /// See if we recognize this DotcomError code
-        ///
-        self = errorDetails.asWCPayChargesError() ?? .otherError(error: error.toAnyError)
-    }
-
-    private static func unwrapError(error: Error) -> WCPayChargesErrorConvertible? {
-        switch error {
-        case let DotcomError.unknown(code, message):
-            return WCPayChargesDotcomErrorDetails(code: code, message: message)
-        case let NetworkError.unacceptableStatusCode(_, response):
-            guard let response,
-                  let errorDetails = try? JSONDecoder().decode(WCPayChargesNetworkErrorDetails.self, from: response) else {
-                return nil
-            }
-            return errorDetails
-        default:
-            return nil
-        }
-    }
-
-    private struct WCPayChargesNetworkErrorDetails: Decodable, WCPayChargesErrorConvertible {
-        let code: WCPayChargesErrorCode
-        let message: String?
-
-        enum CodingKeys: CodingKey {
-            case code
-            case message
-        }
-    }
-
-    private struct WCPayChargesDotcomErrorDetails: WCPayChargesErrorConvertible {
-        // The response JSON differs from NetworkError above:
-        // `"error": "wcpay_get_charge", "message": "Error fetching charge:"`
-        // It's also decoded further up the chain.
-        let code: WCPayChargesErrorCode
-        let message: String?
-
-        init?(code: String, message: String?) {
-            guard let errorCode = WCPayChargesErrorCode(rawValue: code) else {
-                return nil
-            }
-
-            self.code = errorCode
-            self.message = message
-        }
-    }
-
-    public var errorDescription: String? {
-        switch self {
-        case .noSuchChargeError(let message):
-            /// Return the message directly from the store
-            /// "Error: No such charge: 'ch_3KMVapErrorERROR'"
-            return message
-        case .otherError(let error):
-            return error.localizedDescription
-        }
-    }
-}
-
-enum WCPayChargesErrorCode: String, Decodable {
-    case getChargeError = "wcpay_get_charge"
-    case unknown
-}
-
-protocol WCPayChargesErrorConvertible {
-    var code: WCPayChargesErrorCode { get }
-    var message: String? { get }
-}
-
-extension WCPayChargesErrorConvertible {
-    func asWCPayChargesError() -> WCPayChargesError? {
-        switch code {
-        case .getChargeError:
-            guard let message,
-                  message.starts(with: "Error: No such charge") else {
-                return nil
-            }
-            return .noSuchChargeError(message: message)
-        default:
-            return nil
-        }
-    }
-}
